requirements.txt
streamlit
pandas
mysql-connector-python
numpy
altair
plotly
datetime
